+++ examples/hello_world_0.py
@@ -1,6 +1,6 @@
 from random import randint
 
-from gladius import Gladius, Component, EventRequest
+from gladius import Gladius, Component, Event
 from gladius.daisyui import DaisyUI
 
 # glados daisyui
@@ -8,10 +8,10 @@ g = Gladius()
 d = DaisyUI(g)
 
 # callbacks
-async def hello_button_click(button: Component, req: EventRequest):
+async def hello_button_click(button: Component, event: Event):
     hello_text.content = f'Hello {randint(0, 100)}'
 
-async def world_button_click(button: Component, req: EventRequest):
+async def world_button_click(button: Component, event: Event):
     world_text.content = f'World {randint(0, 100)}'
 
 # page